pyproject.toml
[project]
name = "voice-ai-receptionist"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pipecat-ai-flows>=0.0.22",
    "pipecat-ai[cartesia,deepgram,local-smart-turn-v3,openai,runner,silero,webrtc]>=0.0.98",
    "python-dotenv>=1.2.1",
]
